move app.ota_client as otaclient.ota_core
